+++ nilearn/tests/test_niimg_conversions.py
@@ -488,24 +488,30 @@ def test_concat_niimg_dtype():
 
 
 def nifti_generator(buffer):
+    rng = np.random.RandomState(42)
     for i in range(10):
-        buffer.append(Nifti1Image(np.random.random((10, 10, 10)), np.eye(4)))
+        buffer.append(Nifti1Image(rng.random_sample((10, 10, 10)), np.eye(4)))
         yield buffer[-1]
 
 
 def test_iterator_generator():
     # Create a list of random images
-    l = [Nifti1Image(np.random.random((10, 10, 10)), np.eye(4))
-         for i in range(10)]
-    cc = _utils.concat_niimgs(l)
+    rng = np.random.RandomState(42)
+    list_images = [
+        Nifti1Image(
+            rng.random_sample((10, 10, 10)), np.eye(4)
+        )
+        for i in range(10)
+    ]
+    cc = _utils.concat_niimgs(list_images)
     assert cc.shape[-1] == 10
-    assert_array_almost_equal(get_data(cc)[..., 0], get_data(l[0]))
+    assert_array_almost_equal(get_data(cc)[..., 0], get_data(list_images[0]))
 
     # Same with iteration
-    i = image.iter_img(l)
+    i = image.iter_img(list_images)
     cc = _utils.concat_niimgs(i)
     assert cc.shape[-1] == 10
-    assert_array_almost_equal(get_data(cc)[..., 0], get_data(l[0]))
+    assert_array_almost_equal(get_data(cc)[..., 0], get_data(list_images[0]))
 
     # Now, a generator
     b = []